Closes #1 Allow users to configure email address
The /:username/profile route GETs the user profile configuration page,
and POSTing to that will update the profile. Right now the email address
is the only thing in the profile. The user's profile email address is
used to determine where to send warning emails (as opposed to the
address in the configuration file).